BREAKING(json): rename /common to /types
