+++ json/mod.ts
@@ -20,6 +20,6 @@
  */
 
 export * from "./concatenated_json_parse_stream.ts";
-export * from "./common.ts";
+export * from "./types.ts";
 export * from "./json_parse_stream.ts";
 export * from "./json_stringify_stream.ts";
